Added All DAO

+++ project2/shared-libraries/persistentdata/src/main/java/com/revature/daos/BusinessCardDAO.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BusinessCardDAO extends JpaRepository<BusinessCard, Integer> {
+public interface BusinessCardDAO extends JpaRepository<BusinessCard, Long> {
 
 }
